requirements.txt
attrs==21.4.0
click==8.0.3
Flask==2.0.2
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.6
pytest==6.2.5
python-mock==0.0.3
toml==0.10.2
Werkzeug==2.0.2
allure-pytest==2.8.18
pytest==6.0.1
pytest-xdist==2.1.0
requests==2.24.0
allure-pytest==2.8.11
allure-python-commons==2.8.11
attrs==19.3.0
importlib-metadata==1.5.0
more-itertools==8.2.0
packaging==20.1
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.6
pytest==5.3.5
pytest-runner==5.2
six==1.14.0
wcwidth==0.1.8
zipp==3.0.0
mock==4.0.3

